Address PR comments and remove individual timers
